+++ src/store/settingsModule.ts
@@ -1,5 +1,6 @@
 import type { ZBSettings, ZBSettingsFlags } from "@/data/ZBSettings";
 import { loadSettingsFromStore } from "./localStoragePlugin";
+import { changeColorMode } from "@/assets/ColorMode";
 
 // module state object
 const state = JSON.parse(loadSettingsFromStore());
@@ -19,6 +20,13 @@ const mutations = {
 
   changeLanguage(state: ZBSettings, newLanguage: string): void {
     state.language = newLanguage;
+  },
+  changeGridState(state: ZBSettings, newGridState: boolean): void {
+    state.showGrid = newGridState;
+  },
+  changeColorMode(state: ZBSettings, newColorMode: string): void {
+    state.colorMode = newColorMode;
+    changeColorMode(newColorMode);
   },
 };